refactor NavCard + change social icons
