+++ components/NavCard.vue
