+++ components/VsfFooter.vue
@@ -9,36 +9,42 @@
         <div>
           <Logo class="h-auto mb-4 w-44" />
           <h2 class="mb-2 text-sm font-medium">Stay connected:</h2>
-          <div class="flex">
+          <div class="flex gap-1">
             <a
               href="https://github.com/DivanteLtd/vue-storefront"
               target="_blank"
-              ><img
-                src="https://uploads-ssl.webflow.com/5e7cf661c23ac9df156d9c3d/5ed776f6432bf510fc0a18e2_github%202.svg"
-                class="w-auto h-5"
-            /></a>
-            <a href="https://twitter.com/VueStorefront" target="_blank"
-              ><img
-                src="https://uploads-ssl.webflow.com/5e7cf661c23ac9df156d9c3d/5ed7770740bbe258e8da79d0_twitter.svg"
-                class="w-auto h-5 ml-3"
-            /></a>
-            <a href="https://www.youtube.com/vuestorefront" target="_blank"
-              ><img
-                src="https://uploads-ssl.webflow.com/5e7cf661c23ac9df156d9c3d/5ed77728bc0f1e2446a140ee_youtube.svg"
-                class="w-auto h-5 ml-3"
-            /></a>
+              class="hover:text-neutral-900 hover:dark:text-white"
+            >
+              <Icon icon="mdi:github" width="24" class="text-inherit" />
+            </a>
+            <a
+              href="https://twitter.com/VueStorefront"
+              target="_blank"
+              class="hover:text-neutral-900 hover:dark:text-white"
+            >
+              <Icon icon="mdi:twitter" width="24" />
+            </a>
+            <a
+              href="https://www.youtube.com/vuestorefront"
+              target="_blank"
+              class="hover:text-neutral-900 hover:dark:text-white"
+            >
+              <Icon icon="mdi:youtube" width="24" />
+            </a>
             <a
               href="https://www.linkedin.com/company/vue-storefront/"
               target="_blank"
-              ><img
-                src="https://uploads-ssl.webflow.com/5e7cf661c23ac9df156d9c3d/5ed777169d856257c7216ca4_linkedin.svg"
-                class="w-auto h-5 ml-3"
-            /></a>
-            <a href="https://www.facebook.com/vuestorefront/" target="_blank"
-              ><img
-                src="https://uploads-ssl.webflow.com/5e7cf661c23ac9df156d9c3d/5ed7773c6e51067075010d81_facebook.svg"
-                class="w-auto h-5 ml-1"
-            /></a>
+              class="hover:text-neutral-900 hover:dark:text-white"
+            >
+              <Icon icon="mdi:linkedin" width="24" />
+            </a>
+            <a
+              href="https://www.facebook.com/vuestorefront/"
+              target="_blank"
+              class="hover:text-neutral-900 hover:dark:text-white"
+            >
+              <Icon icon="mdi:facebook" width="24" />
+            </a>
           </div>
 
           <h2 class="mt-4 mb-4 text-sm font-medium">Backed by:</h2>
