+++ components/landing/CoreDocsList.vue
@@ -1,6 +1,8 @@
 <template>
   <section class="mt-24 custom-block">
-    <header class="flex items-center justify-between">
+    <header
+      class="flex flex-col justify-between gap-4 md:items-center md:flex-row"
+    >
       <div>
         <h2>New to Vue Storefront?</h2>
         <div class="custom-block subheader">
@@ -10,19 +12,19 @@
           </p>
         </div>
       </div>
-      <div class="flex items-center">
+      <div class="flex items-center shrink-0">
         <a
           href="https://docs.vuestorefront.io/v2/"
           class="inline-block px-4 py-2 text-sm font-medium text-left no-underline !border-b-0 rounded indent-0 group-hover:underline bg-slate-200 text-slate-800 dark:bg-neutral-800 dark:text-slate-200"
         >
-          Go to Core Docs
+          View Core Docs
         </a>
       </div>
     </header>
 
     <div class="grid grid-cols-12 gap-8 mt-8 custom-block">
       <div class="col-span-12 md:col-span-4">
-        <NavigationalCard
+        <NavCard
           to="/guide"
           title="What is Vue Storefront?"
           description="An introduction to Vue Storefront and how it can help you build better storefronts."
@@ -32,7 +34,7 @@
         />
       </div>
       <div class="col-span-12 md:col-span-4">
-        <NavigationalCard
+        <NavCard
           to="/guide"
           title="Vue Storefront Architecture"
           description="Composable commerce lets you freely shape your customer experience."
@@ -42,7 +44,7 @@
         />
       </div>
       <div class="col-span-12 md:col-span-4">
-        <NavigationalCard
+        <NavCard
           to="/guide"
           title="Extending Vue Storefront"
           description="Customize your Vue Storefront project to meet your needs."
@@ -56,10 +58,10 @@
 </template>
 
 <script>
-import NavigationalCard from '../NavigationalCard.vue'
+import NavCard from '../NavCard.vue'
 export default {
   components: {
-    NavigationalCard
+    NavCard
   }
 }
 </script>
